@@ -1,21 +1,20 @@
-package com.myskng.virtualchemlight.Activity
+package com.myskng.virtualchemlight.activity
 
 import android.content.Context
+import androidx.databinding.DataBindingUtil
 import android.media.AudioManager
 import android.media.MediaPlayer
 import android.os.Build
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.VibrationEffect
 import android.os.Vibrator
-import android.support.design.widget.Snackbar
+import com.google.android.material.snackbar.Snackbar
 import android.util.Log
 import android.view.*
-import android.widget.ImageView
-import butterknife.BindView
-import butterknife.ButterKnife
 import com.myskng.virtualchemlight.R
-import com.myskng.virtualchemlight.UO.UOSensor
+import com.myskng.virtualchemlight.uo.UOSensor
+import com.myskng.virtualchemlight.databinding.ActivityUoBinding
 import kotlinx.coroutines.experimental.Job
 import kotlinx.coroutines.experimental.android.UI
 import kotlinx.coroutines.experimental.launch
@@ -23,14 +22,9 @@ import kotlin.coroutines.experimental.suspendCoroutine
 import kotlin.math.absoluteValue
 
 class UOActivity : AppCompatActivity() {
-    @BindView(R.id.UOimageViewMAX)
-    lateinit var uoImageViewMAX: ImageView
-    @BindView(R.id.UOimageViewNormal)
-    lateinit var uoImageViewNormal: ImageView
-    @BindView(R.id.UOimageViewOFF)
-    lateinit var uoImageViewOFF: ImageView
+    private lateinit var binding: ActivityUoBinding
 
-    val parentView by lazy { findViewById<View>(android.R.id.content) }
+    private val parentView by lazy { findViewById<View>(android.R.id.content) }
 
     private lateinit var uoSensor: UOSensor
     private lateinit var uoSound: MediaPlayer
@@ -42,14 +36,14 @@ class UOActivity : AppCompatActivity() {
 
     private val onGestureListener = object : GestureDetector.SimpleOnGestureListener() {
         override fun onFling(e1: MotionEvent?, e2: MotionEvent?, velocityX: Float, velocityY: Float): Boolean {
-            //null check
+            // null check
             if (e1 == null || e2 == null) return super.onFling(e1, e2, velocityX, velocityY)
-            //log
+            // log
             Log.i("UO", "velocityX:${velocityX.absoluteValue}")
             Log.i("UO", "velocityY:${velocityY.absoluteValue}")
             Log.i("UO", "distanceX:${if (e2.x > e1.x) e2.x - e1.x else e1.x - e2.x}")
             Log.i("UO", "distanceY:${if (e2.y > e1.y) e2.y - e1.y else e1.y - e2.y}")
-            //detect left right swipe
+            // detect left right swipe
             val distX = if (e2.x > e1.x) e2.x - e1.x else e1.x - e2.x
             val distY = if (e2.y > e1.y) e2.y - e1.y else e1.y - e2.y
             val detected = when (true) {
@@ -68,24 +62,28 @@ class UOActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_uo)
-        //ButterKnife
-        ButterKnife.bind(this)
-        //Gesture
+        // Binding
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_uo)
+        // Gesture
         gestureDetector = GestureDetector(this, onGestureListener)
-        //Prepare resource
+        // Prepare resource
         vibrator = getSystemService(Context.VIBRATOR_SERVICE) as Vibrator
         uoSound = MediaPlayer.create(this, R.raw.uosound)
         uoSensor = UOSensor(this)
         uoSensor.startSensor()
         uoSensor.onUOIgnition = this::onUOIgnition
-        uoImageViewMAX.alpha = 0f
-        uoImageViewNormal.alpha = 0f
-        //set screen brightness
+        binding.UOimageViewMAX.alpha = 0f
+        binding.UOimageViewNormal.alpha = 0f
+        // UI
+        window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN
+                .or(View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY)
+                .or(View.SYSTEM_UI_FLAG_HIDE_NAVIGATION)
+        // set screen brightness
         val lp = window.attributes
         lp.screenBrightness = WindowManager.LayoutParams.BRIGHTNESS_OVERRIDE_FULL
         window.attributes = lp
         window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        //get volume and show notification
+        // get volume and show notification
         val audioManager = getSystemService(Context.AUDIO_SERVICE) as AudioManager
         val musicVolume = audioManager.getStreamVolume(AudioManager.STREAM_MUSIC)
         if (musicVolume == 0) {
@@ -101,42 +99,42 @@ class UOActivity : AppCompatActivity() {
 
     private fun onUOIgnition() {
         launch(UI, parent = rootJob) {
-            //check if locked
+            // check if locked
             if (uoLock) return@launch
             uoLock = true
-            //stop animation
+            // stop animation
             uoAnimatorList.forEach { it ->
                 it.cancel()
             }
             uoAnimatorList.clear()
-            //reset alpha
-            uoImageViewMAX.alpha = 0f
-            uoImageViewNormal.alpha = 0f
-            //start UO
+            // reset alpha
+            binding.UOimageViewMAX.alpha = 0f
+            binding.UOimageViewNormal.alpha = 0f
+            // start UO
             uoSound.start()
-            //vibrate
+            // vibrate
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 vibrator.vibrate(VibrationEffect.createOneShot(500, 255))
             }
             Log.i("UO", "UO START!!")
-            //uo effect
+            // uo effect
             suspendCoroutine<Unit> {
-                with(uoImageViewMAX.animate().alpha(1.0f)) {
+                with(binding.UOimageViewMAX.animate().alpha(1.0f)) {
                     uoAnimatorList.add(this)
                     this.duration = 500
                     this.withEndAction { it.resume(Unit) }
                 }
             }
             suspendCoroutine<Unit> {
-                uoImageViewNormal.alpha = 1.0f
-                with(uoImageViewMAX.animate().alpha(0.0f)) {
+                binding.UOimageViewNormal.alpha = 1.0f
+                with(binding.UOimageViewMAX.animate().alpha(0.0f)) {
                     uoAnimatorList.add(this)
                     this.duration = 60 * 1000
                     this.withEndAction { it.resume(Unit) }
                 }
             }
             suspendCoroutine<Unit> {
-                with(uoImageViewNormal.animate().alpha(0.0f)) {
+                with(binding.UOimageViewNormal.animate().alpha(0.0f)) {
                     uoAnimatorList.add(this)
                     this.duration = 60 * 2 * 1000
                     this.withEndAction { it.resume(Unit) }
@@ -148,18 +146,18 @@ class UOActivity : AppCompatActivity() {
     }
 
     private fun onUODispose() {
-        //stop animation
+        // stop animation
         uoAnimatorList.forEach { it ->
             it.cancel()
         }
         uoAnimatorList.clear()
-        //reset alpha with animation
-        with(uoImageViewMAX.animate().alpha(0.0f)) {
+        // reset alpha with animation
+        with(binding.UOimageViewMAX.animate().alpha(0.0f)) {
             this.duration = 500
             this.withEndAction { uoLock = false }
             uoAnimatorList.add(this)
         }
-        with(uoImageViewNormal.animate().alpha(0.0f)) {
+        with(binding.UOimageViewNormal.animate().alpha(0.0f)) {
             this.duration = 500
             this.withEndAction { uoLock = false }
             uoAnimatorList.add(this)
