@@ -1,6 +1,6 @@
-package com.myskng.virtualchemlight.View
+package com.myskng.virtualchemlight.view
 
-import android.support.v4.view.PagerAdapter
+import androidx.viewpager.widget.PagerAdapter
 import android.view.View
 import android.view.ViewGroup
 
@@ -13,7 +13,7 @@ class ViewPagerAdapter : PagerAdapter() {
 
     override fun instantiateItem(container: ViewGroup, position: Int): Any {
         val view = PresenterList.get(position).view
-        //workaround 何故か最後のアイテムだけ最初からどこかに所属している状態になるので外してやる
+        // workaround 何故か最後のアイテムだけ最初からどこかに所属している状態になるので外してやる
         (view.parent as ViewGroup?)?.removeView(view)
         container.addView(view)
         return view
